somewhat ready for comp
