+++ src/main/java/frc/robot/commands/mastertoggle/dstoggle/RetractArm.java
@@ -2,12 +2,13 @@ package frc.robot.commands.mastertoggle.dstoggle;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.dreadsubsystem.MasterSubsystem;
+import frc.robot.subsystems.dreadsubsystem.Solenoids;
 
 public class RetractArm extends CommandBase {
-  private final MasterSubsystem m_solenoids;
+  private final Solenoids m_solenoids;
   private boolean m_isDone;
 
-  public RetractArm(MasterSubsystem subsystem) {
+  public RetractArm(Solenoids subsystem) {
     m_solenoids = subsystem;
     m_isDone = false;
     addRequirements(subsystem);
